functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.2.0
fastapi>=0.100.0
pydantic[email]>=2.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.24.1
mangum>=0.17.0
python-multipart>=0.0.6
uvicorn>=0.22.0
